+++ lumin/nn/callbacks/loss_callbacks.py
@@ -12,8 +12,8 @@ class GradClip(Callback):
 
     Arguments:
         clip: value to clip at
-        clip_norm: whether to clip according to norm (`nn.utils.clip_grad_norm_`) or value (`nn.utils.clip_grad_value_`)
-         model: :class:`Model` with parameters to clip gradients, alternatively call :meth:`set_model`
+        clip_norm: whether to clip according to norm (nn.utils.clip_grad_norm_) or value (nn.utils.clip_grad_value_)
+         model: :class:Model with parameters to clip gradients, alternatively call :meth:set_model
 
     Examples::
         >>> grad_clip = GradClip(1e-5)